refactor(backend): make yggdrasil apis use new auth module
