+++ apps/backend/src/yggdrasil/webapis/authserver/invalidate.ts
@@ -1,6 +1,7 @@
 import { Static, t } from 'elysia';
 import { SessionService } from '~backend/services/auth/session';
 import { yggTokenSchema } from '~backend/yggdrasil/yggdrasil.entities';
+import { YggdrasilService } from '~backend/yggdrasil/yggdrasil.service';
 
 export const invalidateBodySchema = yggTokenSchema;
 export const invalidateResponseSchema = t.Void();
@@ -8,5 +9,8 @@ export const invalidateResponseSchema = t.Void();
 export async function invalidate(
 	body: Static<typeof invalidateBodySchema>,
 ): Promise<Static<typeof invalidateResponseSchema>> {
-	await SessionService.invalidate(body.accessToken);
+	const accessToken = YggdrasilService.parseAccessToken(body.accessToken);
+	if (!accessToken) throw new Error('Invalid session!');
+
+	await SessionService.revoke(accessToken.sessionId);
 }
