+++ apps/backend/src/yggdrasil/webapis/authserver/validate.ts
@@ -1,6 +1,8 @@
 import { yggTokenSchema } from '~backend/yggdrasil/yggdrasil.entities';
 import { Static, t } from 'elysia';
-import { SessionScope, SessionService } from '~backend/services/auth/session';
+import { SessionService } from '~backend/services/auth/session';
+import { SessionScope } from '~backend/auth/auth.entities';
+import { YggdrasilService } from '~backend/yggdrasil/yggdrasil.service';
 
 export const validateBodySchema = yggTokenSchema;
 export const validateResponseSchema = t.Void();
@@ -9,10 +11,14 @@ export const validateDataSchema = t.Boolean();
 export async function validate(
 	body: Static<typeof validateBodySchema>,
 ): Promise<Static<typeof validateDataSchema>> {
-	const session = (await SessionService.validate(body.accessToken))?.session;
+	const accessToken = YggdrasilService.parseAccessToken(body.accessToken);
+	if (!accessToken) throw new Error('Invalid session!');
+
+	const session = (await SessionService.validate(accessToken.sessionId, accessToken.sessionToken))
+		?.session;
 	return !(
 		session &&
-		session.scope === SessionScope.YGGDRASIL &&
+		session.metadata.scope === SessionScope.YGGDRASIL &&
 		// When client token is provided, check if it matches, otherwise ignore it.
 		(!body.clientToken || body.clientToken === session.metadata.clientToken)
 	);
