+++ apps/backend/src/yggdrasil/webapis/sessionserver/join-server.ts
@@ -1,6 +1,7 @@
 import { Static, t } from 'elysia';
 import { YggdrasilRepository } from '~backend/yggdrasil/yggdrasil.repository';
 import { SessionService } from '~backend/services/auth/session';
+import { YggdrasilService } from '~backend/yggdrasil/yggdrasil.service';
 
 export const joinServerBodySchema = t.Object({
 	accessToken: t.String(),
@@ -12,7 +13,10 @@ export const joinServerResponseSchema = t.Void();
 export async function joinServer(
 	body: Static<typeof joinServerBodySchema>,
 ): Promise<Static<typeof joinServerResponseSchema>> {
-	const session = await SessionService.validate(body.accessToken);
+	const accessToken = YggdrasilService.parseAccessToken(body.accessToken);
+	if (!accessToken) throw new Error('Invalid session!');
+
+	const session = await SessionService.validate(accessToken.sessionId, accessToken.sessionToken);
 	if (!session) throw new Error('Invalid session!');
 
 	// [TODO] Make this configurable
